+++ integration-tests/zombie-tests/src/main.spec.ts
@@ -1,6 +1,6 @@
 import { expect, describe, it } from "vitest"
 import { start } from "smoldot"
-import { AccountId, createClient } from "@polkadot-api/client"
+import { AccountId, createClient } from "polkadot-api"
 import { getSmProvider } from "@polkadot-api/sm-provider"
 import { WebSocketProvider } from "@polkadot-api/ws-provider/node"
 import { createClient as createRawClient } from "@polkadot-api/substrate-client"